Fixed bugs

+++ Crowdfund.Core/Services/Options/ProjectOptions/CreateProjectOptions.cs
@@ -13,7 +13,7 @@ namespace Crowdfund.Core.Services.Options.ProjectOptions
 
         public DateTime? DueTo { get; set; }
 
-        public decimal Goal { get; set; }
+        public int Goal { get; set; }
         
         public int CategoryId { get; set;}
     }
