+++ Crowdfund.Web/Models/Dashboard/RewardFormOptions.cs
@@ -8,7 +8,7 @@ namespace Crowdfund.Web.Models.Dashboard
 
         public string Description { get; set; }
 
-        public decimal MinAmount { get; set; }
+        public int MinAmount { get; set; }
 
         public int? Quantity { get; set; }
     }
